refactor(web): move submit form to file
